Create mock files
