+++ Tests/ClientTests/RestoreTabManagerTests.swift
@@ -112,45 +112,3 @@ final class RestoreTabManagerTestsTests: XCTestCase {
         return subject
     }
 }
-
-// MARK: - MockRestoreTabManagerDelegate
-final class MockRestoreTabManagerDelegate: RestoreTabManagerDelegate {
-    var needsTabRestoreCalled = 0
-    var needsNewTabOpenedCalled = 0
-
-    func needsTabRestore() {
-        needsTabRestoreCalled += 1
-    }
-
-    func needsNewTabOpened() {
-        needsNewTabOpenedCalled += 1
-    }
-}
-
-// MARK: - MockPresenter
-final class MockPresenter: Presenter {
-    var savedViewControllerToPresent: UIViewController?
-    var savedAnimated: Bool?
-    var savedCompletion: (() -> Void)?
-
-    func present(_ viewControllerToPresent: UIViewController,
-                 animated flag: Bool,
-                 completion: (() -> Void)?) {
-        savedViewControllerToPresent = viewControllerToPresent
-        savedAnimated = flag
-        savedCompletion = completion
-    }
-}
-
-// MARK: - MockRestoreAlertCreator
-final class MockRestoreAlertCreator: RestoreAlertCreator {
-    var savedOkayCallback: (() -> Void)?
-    var savedNoCallback: (() -> Void)?
-
-    func restoreTabsAlert(okayCallback: @escaping () -> Void,
-                          noCallback: @escaping () -> Void) -> UIAlertController {
-        savedOkayCallback = okayCallback
-        savedNoCallback = noCallback
-        return UIAlertController()
-    }
-}
